rfcfold: try to improve sed compatibility
The POSIX standard seems to indicate that that the label names need
to be separated by whitespace from both the 'b' and 't' commands.
The FreeBSD sed man-page indicates that the label name needs to be
separated by space from the 't' command.

The 'b' command without a label branches to the end of the script,
thus there is no need for a specific end label, so remove it.

Use one letter label names to keep total line length below 69 chars.