rfcfold: write return value 255 as 255, not -1
Writing a return value of -1 results in returning 255.  This value
is used as exit code, thus checking the exit code needs to check
for 255, not -1.  The exit code is an integer between 0 and 255, it
cannot be -1.  This is reflected in the test suite, where 255 is used
instead of -1.

Thus write 255 in the code, and call the exit code 255 in the help
text.

This might help in avoiding to confuse users of rfcfold.