rfcfold: tighten lower bound for column to fold on
If no strategy is specified via command line option, an automatic
folding strategy is used.  This autometic folding strategy attempts
to fold the input using strategy one and falls back on using strategy
two if needed.  Thus the minimum value accepted for option "-c" when
the automatic strategy is used needs to allow for the use of folding
strategy two.

The minimum folding column is determined by the header text with a
prefix of three equal signs and a space, and a suffix of a space and
three equal signs.