rfcfold: test all folding columns for space
Strategy 1 cannot correctly fold input with a space character in the
folding column.  If the line that needs folding is sufficiently long,
it needs to be folded more than once.  No folding column may contain
a space character.

Add tests for a space in the first, second, and third folding column.